Move route handlers to separate modules
